Provide Information on Data Availabilities

Data availability of raw/validated/calibrated data for parameter para, measured at monitoring point moniPoint between dateFirst and dateLast.

Usage

hsMiaCsoDataAvailability(
  level,
  moniPoint,
  parName,
  dateFirst = NULL,
  dateLast = NULL,
  tstep = NULL,
  dbg = FALSE
)

Arguments

level

one of "r" (= raw), "v" (= validated), "c" (= calibrated)

moniPoint

one of "STA" (= Stallstr.), "TEG" (= Tegeler Weg), "MUE" (= Muehlendamm)

parName

e.g. "AFS", "CSB", "CSBf", ...

dateFirst

Date object representing first date to be considered

dateLast

Date object representing last date to be considered

tstep

expected time step between time stamps in seconds. Default: minimum time difference found between consecutive timestamps in given interval

dbg

If TRUE, debug messages will be shown

Examples

if (FALSE) {
# Get data availability of raw data of parameter "CSBf", measured at
# monitoring point "TEG" (Tegeler Weg) between 2011-04-01 and 2011-10-01
da <- hsMiaCsoDataAvailability("r", "TEG", "CSBf", "2011-04-01", "2011-10-01")
head(da)
}
# Output:
#   myInterval myCount   myAvail
# 1 2011-04-01    1440 100.00000
# 2 2011-04-02    1440 100.00000
# 3 2011-04-03    1440 100.00000
# 4 2011-04-04    1440 100.00000
# 5 2011-04-05    1440 100.00000
# 6 2011-04-06    1409  97.84722