Plot all MIA CSO Data Availabilities

Plots availability of raw and validated data for different monitoring points and parameters. For each monitoring point a pdf file "hsDataAvailability_<MP>" where <MP> is the acronym of the monitoring point is created in the directory \(strPdfDir\).

Usage

hsPlotAllDataAvailabilities(moniPoints, parNames, dates, pdfDir)

Arguments

moniPoints

vector containing names of monitoring points, e.g. c("TEG", "MUE")

parNames

vector containing names of parameters, e.g. c("AFS", "CSB", "CSBf")

dates

vector containing a list of Date objects

pdfDir

path to output directory to which created pdf-files shall be written.

Details

2012-04-17;HSB;example removed

Examples

if (FALSE) {
# Generate pdf files containing data availability plots for parameters 
# "AFS", "CSB", "CSBf", measured at monitoring points "MUE" (Muehlendamm),
# "TEG" (Tegeler Weg), "STA" (Stallstr.) within two different
# time intervals: 2010-01-01 to 2010-07-01 and  2011-01-01 to 2011-07-01.
hsPlotAllDataAvailabilities(
  c("MUE", "TEG", "STA"), 
  c("AFS", "CSB", "CSBf"),
  as.Date(c("2010-01-01", "2010-07-01", "2011-01-01", "2011-07-01")), 
  tempdir()
)

# Show data availability plots for Muehlendamm in pdf viewer
pdfFile <- file.path(tempdir(), "hsDataAvailability_MUE.pdf")
system(paste(options("pdfviewer"), pdfFile))
}